Moving towards a single supported image, Core CLR.
This change eliminates the Mono image, which was not going to be
officially supported anyway, as it was there as a workaround in case
there was a problem with the CoreCLR image. With this change we're
betting that the CoreCLR will be fully sufficient by the time we go
live.