Refactoring how the images are built.
With this change the images for a particular .NET Core version are all
built at once in a single build. This will allow us to send test code
to the build so we can run tests during the build.

This change also makes it building multiple images faster, as they all
can be built at the same time.